+++ doorstop/core/document.py
@@ -7,7 +7,7 @@ import logging
 
 from doorstop.core.base import BaseValidatable
 from doorstop.core.base import auto_load, auto_save, BaseFileObject
-from doorstop.core.types import ID, Level
+from doorstop.core.types import Prefix, ID, Level
 from doorstop.core.item import Item
 from doorstop import common
 from doorstop.common import DoorstopError, DoorstopWarning
@@ -21,7 +21,7 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
     CONFIG = '.doorstop.yml'
     SKIP = '.doorstop.skip'  # indicates this document should be skipped
 
-    DEFAULT_PREFIX = 'REQ'
+    DEFAULT_PREFIX = Prefix('REQ')
     DEFAULT_SEP = ''
     DEFAULT_DIGITS = 3
 
@@ -121,7 +121,7 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
         sets = data.get('settings', {})
         for key, value in sets.items():
             if key == 'prefix':
-                self._data['prefix'] = value.strip()
+                self._data['prefix'] = Prefix(value)
             elif key == 'sep':
                 self._data['sep'] = value.strip()
             elif key == 'parent':
@@ -141,7 +141,7 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
         sets = {}
         for key, value in self._data.items():
             if key == 'prefix':
-                sets['prefix'] = value
+                sets['prefix'] = str(value)
             elif key == 'sep':
                 sets['sep'] = value
             elif key == 'digits':
@@ -209,7 +209,7 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
     @auto_load
     def prefix(self, value):
         """Set the document's prefix."""
-        self._data['prefix'] = value.strip()
+        self._data['prefix'] = Prefix(value)
         # TODO: should the new prefix be applied to all items?
 
     @property
@@ -309,7 +309,7 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
                         self.path, self.root, identifier,
                         level=nlevel)
         self._items.append(item)
-        if settings.REORDER and level and reorder:
+        if level and reorder:
             self.reorder(keep=item)
         return item
 
@@ -328,25 +328,28 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
         item = self.find_item(identifier)
         item.delete()
         self._items.remove(item)
-        if settings.REORDER and reorder:
+        if reorder:
             self.reorder()
         return item
 
-    def reorder(self, start=None, keep=None):
+    def reorder(self, items=None, start=None, keep=None):
         """Reorder a document's items.
 
+        @param items: items to reorder (None = reorder instance items)
         @param start: level to start numbering (None = use current start)
         @param keep: item or ID to keep over duplicates
 
         """
+        items = items or self.items
         keep = self.find_item(keep) if keep else None
         logging.info("reordering {}...".format(self))
-        self._reorder(self.items, start=start, keep=keep)
+        self._reorder(items, start=start, keep=keep)
 
     @staticmethod
     def _reorder(items, start=None, keep=None):
         """Reorder a document's items.
 
+        @param items: items to reorder
         @param start: level to start numbering (None = use current start)
         @param keep: item to keep over duplicates
 
@@ -446,11 +449,15 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
         assert kwargs.get('document_hook') is None
         hook = item_hook if item_hook else lambda **kwargs: []
         logging.info("checking document {}...".format(self))
-        # Reorder levels
+        # Check for items
+        items = self.items
+        if not items:
+            return DoorstopWarning("no items")
+        # Reorder or check item levels
         if settings.REORDER:
-            self.reorder()
-        # Check levels
-        yield from self._get_issues_level()
+            self.reorder(items=items)
+        elif settings.CHECK_LEVELS:
+            yield from self._get_issues_level(items)
         # Check each item
         for item in self:
             # Check item
@@ -460,14 +467,10 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
                 if isinstance(issue, Exception):
                     yield type(issue)("{}: {}".format(item.id, issue))
 
-    def _get_issues_level(self, items=None):
+    @staticmethod
+    def _get_issues_level(items):
         """Yield all the document's issues related to item level."""
-        items = items or self.items
-        # Check for items
-        if not items:
-            return DoorstopWarning("no items")
-        # Check item levels
-        prev = items[0]
+        prev = items[0] if items else None
         for item in items[1:]:
             pid = prev.id
             plev = prev.level
@@ -501,11 +504,3 @@ class Document(BaseValidatable, BaseFileObject):  # pylint: disable=R0902,R0904
         for item in self:
             item.delete()
         super().delete(self.config)
-
-
-# attribute formatters #######################################################
-
-# TODO: move this code to calling locations
-def get_prefix(value):
-    """Get a prefix from a document or string."""
-    return str(value).split(' ')[0]