+++ doorstop/core/test/test_document.py
@@ -209,7 +209,7 @@ class TestDocument(unittest.TestCase):  # pylint: disable=R0904
         """Verify an item can be added to an new document."""
         document = MockDocument(NEW, ROOT)
         document.prefix = 'NEW'
-        self.assertIsNot(None, document.add_item())
+        self.assertIsNot(None, document.add_item(reorder=False))
         mock_new.assert_called_once_with(None, document,
                                          NEW, ROOT, 'NEW001',
                                          level=None)
@@ -218,7 +218,7 @@ class TestDocument(unittest.TestCase):  # pylint: disable=R0904
         """Verify an added item is contained in the document."""
         item = self.document.items[0]
         self.assertIn(item, self.document)
-        item2 = self.document.add_item()
+        item2 = self.document.add_item(reorder=False)
         self.assertIn(item2, self.document)
 
     @patch('doorstop.core.document.Document._reorder')
@@ -236,9 +236,18 @@ class TestDocument(unittest.TestCase):  # pylint: disable=R0904
         """Verify a removed item is not contained in the document."""
         item = self.document.items[0]
         self.assertIn(item, self.document)
-        removed_item = self.document.remove_item(item.id)
+        removed_item = self.document.remove_item(item.id, reorder=False)
         self.assertEqual(item, removed_item)
         self.assertNotIn(item, self.document)
+        mock_remove.assert_called_once_with(item.path)
+
+    @patch('os.remove')
+    def test_remove_item_by_item(self, mock_remove):
+        """Verify an item can be removed (by item)."""
+        item = self.document.items[0]
+        self.assertIn(item, self.document)
+        removed_item = self.document.remove_item(item, reorder=False)
+        self.assertEqual(item, removed_item)
         mock_remove.assert_called_once_with(item.path)
 
     def test_reorder(self):
@@ -312,15 +321,6 @@ class TestDocument(unittest.TestCase):  # pylint: disable=R0904
         actual = [item.level for item in mock_items]
         self.assertListEqual(expected, actual)
 
-    @patch('os.remove')
-    def test_remove_item_by_item(self, mock_remove):
-        """Verify an item can be removed (by item)."""
-        item = self.document.items[0]
-        self.assertIn(item, self.document)
-        removed_item = self.document.remove_item(item)
-        self.assertEqual(item, removed_item)
-        mock_remove.assert_called_once_with(item.path)
-
     def test_find_item(self):
         """Verify an item can be found by ID."""
         item = self.document.find_item('req2')
@@ -346,7 +346,7 @@ class TestDocument(unittest.TestCase):  # pylint: disable=R0904
         mock_get_issues.return_value = [DoorstopInfo('i')]
         with patch('doorstop.settings.REORDER', True):
             self.assertTrue(self.document.validate())
-        mock_reorder.assert_called_once_with()
+        mock_reorder.assert_called_once_with(items=self.document.items)
         self.assertEqual(5, mock_get_issues.call_count)
 
     @patch('doorstop.core.item.Item.get_issues',